Prefer size_t when expressing an object size in memory

+++ lib/data_node.cc
@@ -70,7 +70,7 @@ static void ProcessPkWareUnixField(DataNode* const node,
   gid_t gid;
   dev_t dev;
   const char* link;
-  uint16_t link_len;
+  size_t link_len;
   if (!ExtraField::parsePkWareUnixField(len, field, node->mode, mt, at, uid,
                                         gid, dev, link, link_len)) {
     return;
